modifications for fsl 3.0.0 and add header
Freescale kernel 3.0.35, rev 3.0.0 requires that mx6_cpu_regulator_init()
be called within the board init function.  Not doing so results in
unrecoverable kernel errors during kernel initialization in calls to
regulator_set_voltage().  Calls to this function seem to be removed
in rev 1.1.0 and later versions of the FSL kernel.

In addition, add the header to the file giving the original authors credit.

Include 'cpu_op-mx6.h' because this was commented out for some reason.

Upstream-Status: Pending

Signed-off-by: John Weber <rjohnweber@gmail.com>